// src/app/user.interface.ts
export interface User {
    readonly id: number;
    readonly name: string;
    readonly username: string;
    readonly email: string;
    readonly address: {
        readonly street: string;
        readonly suite: string;
        readonly city: string;
        readonly zipcode: string;
        readonly geo: {
            readonly lat: string;
            readonly lng: string;
      };
    };
    addrstr?: string;
    readonly phone: string;
    readonly website: string;
    readonly company: {
        readonly name: string;
        readonly catchPhrase: string;
        readonly bs: string;
    };
  }

  export type Users = User[];